requirements.txt
hmmlearn>=0.3.0
lightgbm>=4.0.0
alpaca-py>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
ta>=0.10.2
plotly>=5.22.0
streamlit>=1.35.0
pyyaml>=6.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0
rich>=13.0.0
exchange-calendars>=4.2.0
pytz>=2023.3
pytest>=7.3.0
pytest-asyncio>=0.21.0
joblib>=1.2.0
